Feature: Namespaces\UnusedUseStatement - support for use groups + fixes
Adds support for use groups, various fixes on detecting if imported
class/function/const is used in the code.

The same alias can be used for class, function and constant,
so we need to detect what is the context of the string token.

For classes and function PHP is case insensitive, but not for constants.
So we can have defined/imported constants with names: `FOO`, `Foo`, `foo`
and all these three are different.