Hotfix: allow null value for properties with nullable type declaration
Since PHP 7.4 we can define property type. In that case, when property
is defined with nullable type we can have default null value, as it is
different from not being initialised.

Removed also invalid test case with callable property type, as it is not
valid in PHP 7.4.

Fixes #72

Also adds new code type for null values for properties with not-nullable
type declaration. This is technically invalid code.